+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
  - Fusion: A null value sends an empty string
+ - Fusion: Removed Audio and USB Breakaway mapping sigs
 
 ## [5.2.3] - 2019-12-03
 ### Added